Rename unblob_check to unblobCheck and move it to dcr.util.encode.
